Fixed issues with PR
